feat: Criação de entidades e repositórios, com a implementação de query native para busca dos dados com filros opcionais

+++ src/scripts_bd/schema.sql
@@ -1,18 +1,18 @@
 CREATE TABLE solicitante (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     nome VARCHAR(255) NOT NULL,
     cpf_cnpj VARCHAR(18) UNIQUE NOT NULL
 );
 
 CREATE TABLE categoria (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     nome VARCHAR(100) NOT NULL
 );
 
 CREATE TABLE solicitacao (
-    id SERIAL PRIMARY KEY,
-    solicitante_id INTEGER NOT NULL,
-    categoria_id INTEGER NOT NULL,
+    id BIGSERIAL PRIMARY KEY,
+    solicitante_id BIGINT NOT NULL,
+    categoria_id BIGINT NOT NULL,
     descricao TEXT NOT NULL,
     valor DECIMAL(10, 2) NOT NULL,
     data_solicitacao TIMESTAMP NOT NULL,
